+++ spec/integration/resource/api500/c7000/fc_network/update_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API500::C7000::FCNetwork, integration: true, type: UPDATE do
+  include_examples 'FCNetworkUpdateExample', 'integration api500 context' do
+    let(:current_client) { $client_500 }
+  end
+
+  include_examples 'ScopeHelperMethodsExample', OneviewSDK::API500::C7000::Scope do
+    let(:current_client) { $client_500 }
+    let(:item) { described_class.get_all(current_client).first }
+  end
+end