+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Unreleased Changes
+
 ## Suggested release: v4.3.0
 
 #### Features supported
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
    - FCoE network
+   - Scopes
 
 # v4.2.0