+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Features supported
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
+   - FC network
    - Logical switch group
    - Scopes