+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Scopes
    - Switches
 
+#### Breaking changes
+1. In the Switch resource, the `statistics` method was fixed, a non-existent endpoint was removed `/rest/switches/statistics/{portName}/subport/{subportNumber}`. Now the method only get statistics of a switch or a specific port.
+
 # v4.2.0
 
 #### New Resources: