+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
    - Ethernet network
    - FC network
+   - FCoE network
    - Logical switch group
+   - Network set
    - Scopes
 
 # v4.2.0