+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Unreleased Changes
 ##### Suggested release: v4.3.0
 
+#### New Features:
+- Added SCMB module and CLI command
+
 #### Bug fixes & Enhancements:
 - [#222](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/222) Error listing the OS Deployment Plans from OneView
-- Added SCMB module and CLI command
 
 ## v4.2.0